pyproject.toml
[project]
name = "urpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach==6.2.0",
    "channels==4.2.0",
    "channels-redis==4.2.1",
    "daphne==4.1.2",
    "django==5.1.7",
    "django-bootstrap5==25.1",
    "django-cachalot==2.7.0",
    "django-compressor==4.5.1",
    "django-extensions==3.2.3",
    "django-htmx==1.23.0",
    "django-libsass==0.9",
    "django-registration==5.1.0",
    "django-reversion==5.1.0",
    "django-transmeta",
    "django-unfold==0.53.0",
    "hypercorn>=0.17.3",
    "markdown==3.7",
    "pillow==11.1.0",
    "psycopg==3.2.6",
    "pymemcache==4.0.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sorl-thumbnail==12.11.0",
    "uvicorn==0.34.0",
    "weasyprint==65.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=5.1.0",
    "ipython>=9.1.0",
]

[tool.uv.sources]
django-transmeta = { git = "https://github.com/SvenBroeckling/django-transmeta.git" }
